+++ helpers/helpers-init.php
@@ -11,3 +11,11 @@ if(get_option('w4os_provide_economy_helpers') &! empty(W4OS_GRID_INFO['economy']
     die();
   }
 }
+
+if ( get_option('w4os_provide_offline_messages') &! empty(W4OS_GRID_INFO['message']) ) {
+  $message = parse_url(W4OS_GRID_INFO['message'])['path'];
+  $url = getenv('REDIRECT_URL');
+  if(preg_match(":^$message/(SaveMessage|RetrieveMessages|offlineim)/:", "$url/")) {
+    require('offline.php');
+  }
+}